+++ src/tests/jdn.rs
@@ -1,4 +1,4 @@
-use crate::{Calendar, DateError, Jdnum, Month};
+use crate::{errors::DateError, Calendar, Jdnum, Month};
 use rstest::rstest;
 use rstest_reuse::{apply, template};